Added ability to transfer from one account to another.
NOTES: need to fix transaction processing so that nothing gets committed until all changes are successful.

(most of the time transaction stuff like this is built in to the DB with various locking mechanisms and rollback and everything. I probably went overboard on this exercise in trying to reinvent the wheel -- out of scope)